+++ convey/__init__.py
@@ -1,3 +1,4 @@
 from .decorators import PickMethod, PickInput
 
 __all__ = ["PickMethod", "PickInput"]
+__version__  = "1.3.16" # to be changed in controller.py too